backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0
email-validator==2.2.0
google-genai>=1.0.0
